requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
PyYAML>=6.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
yfinance>=0.2.0
openpyxl>=3.1.0

